requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-log-python-sdk==0.8.8
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.2.0
click==8.1.7
ConfigArgParse==1.7
cycler==0.11.0
dateparser==1.1.8
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.9.0
filelock==3.12.2
Flask==2.2.5
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.38.0
frozenlist==1.3.3
gevent==22.10.2
geventhttpclient==2.0.10
greenlet==2.0.2
gremlinpython==3.7.0
grpcio==1.58.0
httmock==1.4.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
influxdb-client==1.37.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
llvmlite==0.39.1
locust==2.16.1
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.5.3
msgpack==1.0.5
multidict==6.0.4
Naked==0.1.32
nest-asyncio==1.5.7
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
orjson==3.9.7
packaging==23.1
pandas==1.3.5
patsy==0.5.3
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
prometheus-api-client==0.5.1
protobuf==3.20.3
psutil==5.9.5
pycryptodome==3.4.3
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.6.3
reactivex==4.0.4
redis==5.0.0
regex==2023.8.8
requests==2.31.0
roundrobin==0.0.4
schedule==1.2.0
scikit-learn==1.0.2
scipy==1.7.3
shellescape==3.8.1
six==1.16.0
sklearn==0.0.post9
statsmodels==0.13.5
tenacity==8.2.3
threadpoolctl==3.1.0
tigramite==5.0.0.3
torch==1.13.1
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.2.3
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
